requirements.txt
python-dotenv
asyncpg
sqlalchemy
uvicorn
fastapi
passlib
pyjwt
python-multipart
pydantic
pydantic-settings
